Move CasperPacketHandler to protocol folder
